+++ brawlstats/errors.py
@@ -5,24 +5,6 @@ class RequestError(Exception):
         pass
 
 
-class BadRequest(RequestError):
-    """Raised if you sent a bad request to the API."""
-
-    def __init__(self, url, code):
-        self.code = code
-        self.error = 'You sent a bad request the API.\nURL: ' + url
-        super().__init__(self.code, self.error)
-
-
-class NotFoundError(RequestError):
-    """Raised if the tag was not found."""
-
-    def __init__(self, url, code):
-        self.code = code
-        self.error = 'The tag you entered was not found.\nURL: ' + url
-        super().__init__(self.code, self.error)
-
-
 class Unauthorized(RequestError):
     """Raised if your API Key is invalid or blocked."""